añadir fondos more y my books
