+++ app/src/main/java/com/example/myfavbook/activities/MyBooks.java
@@ -1,0 +1,63 @@
+package com.example.myfavbook.activities;
+
+import android.content.Intent;
+import android.os.Bundle;
+
+import com.example.myfavbook.R;
+import com.example.myfavbook.credentials.Home;
+import com.google.android.material.bottomnavigation.BottomNavigationView;
+import com.google.android.material.snackbar.Snackbar;
+
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.view.MenuItem;
+import android.view.View;
+
+import androidx.navigation.NavController;
+import androidx.navigation.Navigation;
+import androidx.navigation.ui.AppBarConfiguration;
+import androidx.navigation.ui.NavigationUI;
+
+
+
+public class MyBooks extends AppCompatActivity {
+
+
+
+
+    BottomNavigationView bottomNavigationView;
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        bottomNavigationView = findViewById(R.id.btn_navigator);
+        bottomNavigationView.setSelectedItemId(R.id.search);
+        bottomNavigationView.setOnNavigationItemSelectedListener(new BottomNavigationView.OnNavigationItemSelectedListener() {
+            @Override
+            public boolean onNavigationItemSelected(@NonNull MenuItem item) {
+                switch (item.getItemId()) {
+                    case R.id.my_books:
+
+                        return true;
+
+                    case R.id.more:
+                        startActivity(new Intent(getApplicationContext(), More.class));
+                        overridePendingTransition(0, 0);
+                        return true;
+
+                    case R.id.home:
+                        startActivity(new Intent(getApplicationContext(), Home.class));
+                        overridePendingTransition(0, 0);
+                        return true;
+
+                    case R.id.search:
+                        startActivity(new Intent(getApplicationContext(), Search.class));
+                        overridePendingTransition(0, 0);
+                        return true;
+                }
+                return false;
+            }
+        });
+
+
+    }}
